# Lokata.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lokata.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Approach": {
      "get": {
        "tags": [
          "Approach"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Approach"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Approach"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Approach"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Approach"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Approach/{id}": {
      "get": {
        "tags": [
          "Approach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Approach"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approach"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approach"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Approach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Approach"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Approach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category": {
      "get": {
        "tags": [
          "Category"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Competition": {
      "get": {
        "tags": [
          "Competition"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competition"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competition"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competition"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competition"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Competition/{id}": {
      "get": {
        "tags": [
          "Competition"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Competition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competition"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Competition"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Competition"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Competitor": {
      "get": {
        "tags": [
          "Competitor"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competitor"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competitor"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competitor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competitor"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Competitor/{id}": {
      "get": {
        "tags": [
          "Competitor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Competitor"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competitor"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competitor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Competitor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Competitor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Competitor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Approach": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitionsId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "instructorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "umpireId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competition": {
            "$ref": "#/components/schemas/Competition"
          },
          "competitions": {
            "$ref": "#/components/schemas/Competitions"
          },
          "instructor": {
            "$ref": "#/components/schemas/Instructor"
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          },
          "umpire": {
            "$ref": "#/components/schemas/Umpire"
          }
        },
        "additionalProperties": false
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "minAgeInYears": {
            "type": "integer",
            "format": "int32"
          },
          "maxAgeInYears": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "forDisabled": {
            "type": "boolean"
          },
          "forProfessional": {
            "type": "boolean"
          },
          "forFemale": {
            "type": "boolean"
          },
          "forMale": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Competition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "seriesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPointsPerSeries": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Competitions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "additionalProperties": false
      },
      "Competitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDisabledPerson": {
            "type": "boolean",
            "nullable": true
          },
          "professional": {
            "type": "boolean",
            "nullable": true
          },
          "sexId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Instructor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "documentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "document": {
            "$ref": "#/components/schemas/Document"
          },
          "documentValidUntil": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ScoreCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
...
      "TakePlace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tookPlace": {
            "type": "boolean",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}